requirements.txt
asgiref==3.8.1
Django==5.2.3
o==0.0.2b2
pillow==11.3.0
psycopg2==2.9.10
sqlparse==0.5.3
